From Bob Kuehne, warning fixes for OSX build. From Robert Osfield, warnings fixes for Linux build.
git-svn-id: http://www.openscenegraph.org/svn/osg/OpenThreads/trunk@4059 16af8721-9629-0410-8352-f15c8da7e697

+++ examples_src/simplethreader/SimpleThreader.cpp
@@ -94,7 +94,7 @@ public:
 	bar.block(GLOBAL_NUM_THREADS);  // Sync the threads.
 
 	char tmp[80];
-	sprintf(tmp, "StackSize: %d\n", getStackSize());
+	sprintf(tmp, "StackSize: %d\n", static_cast<int>(getStackSize()));
 
         notifyObserversStarted(getThreadId());
         notifyObserversMessage(getThreadId(), "This is a thread message.");
@@ -194,7 +194,7 @@ int main(int argc, char **argv) {
     std::cin >> val;
 
     // Notify the threads to quit, wait for this to happen.
-    for(i=0;i<threads.size();++i) {
+    for(i=0;i<static_cast<int>(threads.size());++i) {
         MyThread *thread = threads[i];
         thread->quit();
     }
@@ -214,7 +214,7 @@ int main(int argc, char **argv) {
     std::cout << std::endl;
 
     // Delete all the threads.
-    for(i=0;i<threads.size();++i) {
+    for(i=0;i<static_cast<int>(threads.size());++i) {
         MyThread *thread = threads[i];
         delete thread;
     }
